Begin support for relaying media
Problems:
When two photos are sent, they get relayed as an album on telegram. Each
photo in the album has its own message ID. This breaks the "forward_key"
ID that we are using to find messages. The album does have a grouped_id
field (see https://t.me/c/1258001846/112488). Too much effort to deal
with this right now.